pyproject.toml
[project]
name = "scholarly-website-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asgiref==3.8.1",
    "async-timeout==5.0.1",
    "black==25.1.0",
    "click==8.1.8",
    "django==5.1.6",
    "django-cors-headers>=4.0.0",
    "django-picklefield==3.2",
    "django-q2==1.7.6",
    "django-redis==5.4.0",
    "djangorestframework==3.15.2",
    "dspy-ai>=2.5.0",
    "gunicorn==23.0.0",
    "isort==6.0.0",
    "mypy-extensions==1.0.0",
    "openai>=1.0.0",
    "packaging==24.2",
    "pathspec==0.12.1",
    "pdfplumber>=0.10.0",
    "pinecone==7.0.2",
    "pinecone-client>=3.0.0",
    "platformdirs==4.3.6",
    "psycopg2-binary==2.9.10",
    "python-dotenv==1.0.1",
    "redis==5.2.1",
    "sqlparse==0.5.3",
    "tomli==2.2.1",
    "typing-extensions==4.12.2",
    "whitenoise==6.9.0",
]
